# openapi.yaml
openapi: 3.0.3
info:
  title: EmgR OpenApi
  description: |
    EmgR app. More at [official doc](https://vaam-store.github.io/image-resizer)
  contact:
    email: selastlambou@gmail.com
    name: Stephane Segning
  license:
    name: MIT
  version: 0.1.2
externalDocs:
  description: |
    Find out more about Managing your Course from afar. Let's meet on
    our website 😉.
  url: https://blog.ssegning.com
servers:
  - url: 'https://emgr.test'
    description: Production server
  - url: 'http://localhost:3000'
    description: Development server
  - url: '{scheme}://{host}:{port}'
    description: Custom server
    variables:
      scheme:
        enum:
          - http
          - https
        default: http
      host:
        default: localhost
      port:
        default: '3000'

tags:
  - name: Images
    description: All about image transorfmation
paths:
  ##########################################################################
  # COURSES
  ##########################################################################
  /api/images/resize:
    get:
      summary: Resize an image
      operationId: resize
      tags:
        - Images
      parameters:
        - $ref: '#/components/parameters/url'
        - $ref: '#/components/parameters/width'
        - $ref: '#/components/parameters/height'
        - $ref: '#/components/parameters/format'
        - $ref: '#/components/parameters/blur_sigma'
        - $ref: '#/components/parameters/grayscale'
      responses:
        '301':
          description: The image was resize and in the location you'll get the link to it
          headers:
            Location:
              description: URI where the image can be downloaded
              schema:
                type: string
                format: uri
        '400':
          description: |
            The request was invalid - e.g. the source URL points at something
            that isn't a decodable image, or exceeds the configured size
            limit. Never a redirect to the caller-supplied URL (see #25).
          headers:
            Cache-Control:
              description: Always `no-store`; error responses are never cached.
              schema:
                type: string
                example: "no-store"
          content:
            text/plain:
              schema:
                type: string
        '502':
          description: |
            The origin server for the requested image failed (unreachable,
            non-2xx, or the connection dropped mid-transfer).
          headers:
            Cache-Control:
              description: Always `no-store`; error responses are never cached.
              schema:
                type: string
                example: "no-store"
          content:
            text/plain:
              schema:
                type: string
        '503':
          description: |
            The service is shedding load (download/processing concurrency
            limits reached) and cannot handle the request right now.
          headers:
            Cache-Control:
              description: Always `no-store`; error responses are never cached.
              schema:
                type: string
                example: "no-store"
          content:
            text/plain:
              schema:
                type: string
  /api/images/files/{key}:
    get:
      summary: Resize an image
      operationId: download
      tags:
        - Images
      parameters:
        - $ref: '#/components/parameters/key'
      responses:
        '200':
          description: Operation performed successfully.
          headers:
            Cache-Control:
              description: Cache control header
              schema:
                type: string
                example: "public, max-age=31536000, immutable"
          content:
            image/png:
              schema:
                type: string
                format: binary
            image/jpeg:
              schema:
                type: string
                format: binary
            image/webp:
              schema:
                type: string
                format: binary
            application/octet-stream:
              schema:
                type: string
                format: binary
        '404':
          description: No image exists for the given key.
          headers:
            Cache-Control:
              description: Always `no-store`; error responses are never cached.
              schema:
                type: string
                example: "no-store"
          content:
            text/plain:
              schema:
                type: string
        '502':
          description: The storage backend failed to serve the image.
          headers:
            Cache-Control:
              description: Always `no-store`; error responses are never cached.
              schema:
                type: string
                example: "no-store"
          content:
            text/plain:
              schema:
                type: string

components:


  ##########################################################################
  # Params
  ##########################################################################
  parameters:
    url:
      name: url
      in: query
      required: true
      description: The url of the image to be resized
      schema:
        $ref: '#/components/schemas/Url'
    width:
      name: width
      in: query
      required: false
      description: The width of the final image
      schema:
        $ref: '#/components/schemas/Size'
    height:
      name: height
      in: query
      required: false
      description: The height of the final image
      schema:
        $ref: '#/components/schemas/Size'
    blur_sigma:
      name: blur_sigma
      in: query
      required: false
      description: How deep the image should be blured
      schema:
        $ref: '#/components/schemas/BlurSigma'
    grayscale:
      name: grayscale
      in: query
      required: false
      description: Should the image be in grayscale?
      schema:
        $ref: '#/components/schemas/Grayscale'
    format:
      name: format
      in: query
      required: false
      description: The format of the final image
      schema:
        $ref: '#/components/schemas/ImageFormat'
    key:
      name: key
      in: path
      required: true
      description: The unique key of the final image
      schema:
        $ref: '#/components/schemas/Key'


  ##########################################################################
  # SCHEMAS (Data Models)
  ##########################################################################
  schemas:
    Key:
      type: string
      # Exactly what CacheService::generate_key produces: an optional
      # STORAGE_SUB_PATH prefix followed by a 64-char lowercase hex SHA-256
      # digest and one of the image extensions the resizer emits. Anything
      # else (absolute paths, traversal, arbitrary strings) is rejected
      # before it reaches any storage backend (#23).
      pattern: '^.*[0-9a-f]{64}\.(jpg|png|webp)$'
    Url:
      type: string
      format: uri
    Size:
      type: integer
      format: int32
      default: 200
      maximum: 4096
      minimum: 10
    BlurSigma:
      type: number
      format: float
      default: 5
      maximum: 100
      minimum: 0
    Grayscale:
      type: boolean
    ImageFormat:
      type: string
      default: jpg
      enum:
        - png
        - webp
        - jpg